requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
deprecation==2.1.0
ecdsa==0.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-GraphQL==2.0.1
graphql-core==2.3.2
graphql-server-core==1.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
promise==2.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.3.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
Rx==1.6.3
six==1.16.0
stripe==6.5.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
